Refactor data dictionary access into class
